requirements.txt
asgiref @ file:///tmp/build/80754af9/asgiref_1625643473416/work
certifi==2021.10.8
Django @ file:///tmp/build/80754af9/django_1625585912945/work
django-ckeditor==6.2.0
django-js-asset==1.2.2
Faker==10.0.0
psycopg2 @ file:///opt/concourse/worker/volumes/live/7541fe58-2a45-4073-5d69-313e733fc729/volume/psycopg2_1612298592535/work
python-dateutil==2.8.2
pytz==2021.3
six==1.16.0
sqlparse @ file:///tmp/build/80754af9/sqlparse_1602184451250/work
text-unidecode==1.3
typing-extensions @ file:///tmp/build/80754af9/typing_extensions_1631814937681/work
